+++ Tests/FO-DICOM.Tests/Network/DicomPresentationContextCollectionTests.cs
@@ -156,5 +156,149 @@ namespace FellowOakDicom.Tests.Network
             Assert.False(presentationContext.HasTransferSyntax(DicomTransferSyntax.ImplicitVRLittleEndian));
         }
 
+        [Fact]
+        public void Add_SamePresentationContextTwice_ShouldSilentlyNotAddTheSecondOne()
+        {
+            // Arrange
+            var presentationContexts = new DicomPresentationContextCollection();
+
+            // Act
+            // 1
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // 2 (extra transfer syntax)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1, DicomTransferSyntax.JPEG2000Lossless }
+            );
+
+            // 3 (user role is true)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: true,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // 4 (provider role is false)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: false,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // 1b (this one is identical to 1 and should not be added)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // Assert
+            Print(presentationContexts);
+            Assert.Equal(4, presentationContexts.Count);
+
+            var pc1 = presentationContexts.Single(pc => pc.UserRole == null && pc.ProviderRole == null && pc.GetTransferSyntaxes().Count == 1);
+            var pc2 = presentationContexts.Single(pc => pc.GetTransferSyntaxes().Count == 2);
+            var pc3 = presentationContexts.Single(pc => pc.UserRole == true);
+            var pc4 = presentationContexts.Single(pc => pc.ProviderRole == false);
+
+            Assert.NotNull(pc1);
+            Assert.NotNull(pc2);
+            Assert.NotNull(pc3);
+            Assert.NotNull(pc4);
+        }
+
+        [Fact]
+        public void Add_SamePresentationContextTwice_ShouldMaintainCorrectPresentationContextIds()
+        {
+            // Arrange
+            var presentationContexts = new DicomPresentationContextCollection();
+
+            // Act
+            // 1
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // 1b (this one is identical to 1 and should not be added)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // 2 (extra transfer syntax)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1, DicomTransferSyntax.JPEG2000Lossless }
+            );
+
+            // 1c (this one is identical to 1 and should not be added)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // 3 (user role is true)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: true,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // 1d (this one is identical to 1 and should not be added)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: null,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // 4 (provider role is false)
+            presentationContexts.Add(
+                abstractSyntax: DicomUID.CTImageStorage,
+                userRole: null,
+                providerRole: false,
+                transferSyntaxes: new[] { DicomTransferSyntax.JPEGProcess14SV1 }
+            );
+
+            // Assert
+            Print(presentationContexts);
+            Assert.Equal(4, presentationContexts.Count);
+
+            var pc1 = presentationContexts.Single(pc => pc.UserRole == null && pc.ProviderRole == null && pc.GetTransferSyntaxes().Count == 1);
+            var pc2 = presentationContexts.Single(pc => pc.GetTransferSyntaxes().Count == 2);
+            var pc3 = presentationContexts.Single(pc => pc.UserRole == true);
+            var pc4 = presentationContexts.Single(pc => pc.ProviderRole == false);
+
+            Assert.NotNull(pc1);
+            Assert.NotNull(pc2);
+            Assert.NotNull(pc3);
+            Assert.NotNull(pc4);
+            Assert.Equal((byte) 1, pc1.ID);
+            Assert.Equal((byte) 3, pc2.ID);
+            Assert.Equal((byte) 5, pc3.ID);
+            Assert.Equal((byte) 7, pc4.ID);
+        }
+
     }
 }