+++ ChangeLog.md
@@ -16,7 +16,8 @@
 * Fix issue where a DICOM server could stop accepting incoming connections if MaxClientsAllowed is configured and one or more connections never close (#1468)
 * Fix issue where a DICOM server could leak memory when one or more connections never close (#1594)
 * Fix the issue of 'DicomAttribute not generated in XML when element is of type DicomFragmentSequence'
- 
+* Prevent adding duplicate presentation contexts to an association request (#1596)
+* Fix issue with missing known DicomTransferSyntax from static DicomTransferSyntax.Entries dictionary (#1644)
 
 #### 5.1.1 (2023-05-29)
 * Fix issue where DicomClient did not send requests when Async Ops Invoked was zero (#1597)